Improve wording of when rpi-update should be run
rpi-update should only be run when you know its going to fix the issue because it can break things. The error will only happen on the raspberry pi 4 with images older than around July of 2019. This should not be presented as a something to try but a fix for a specific issue.